+++ familyhelper-ui-view/packages/component-api/src/api/settingrepo/reset.js
@@ -1,7 +1,0 @@
-// noinspection NpmUsedModulesInstalled,JSUnusedGlobalSymbols
-
-import {post} from '@/util/http';
-
-export function resetFormat() {
-    return post('settingrepo', 'reset-format/', {});
-}