+++ familyhelper-ui-view/packages/webapp-pc/src/elements/modules/iconfont/index.js
@@ -1,2 +1,0 @@
-import './global.css';
-import '@dwarfeng/familyhelper-ui-component-iconfont/src/iconfont.css';