+++ familyhelper-ui-view/packages/webapp-pc/src/elements/modules/overlayScrollBars/index.js
@@ -1,8 +1,0 @@
-import Vue from 'vue';
-
-import { OverlayScrollbarsPlugin } from 'overlayscrollbars-vue';
-import 'overlayscrollbars/css/OverlayScrollbars.css';
-
-import './global.css';
-
-Vue.use(OverlayScrollbarsPlugin);