+++ familyhelper-ui-view/packages/webapp-pc/src/elements/modules/vueCropper/index.js
@@ -1,5 +1,0 @@
-import Vue from 'vue';
-
-import VueCropper from 'vue-cropper';
-
-Vue.use(VueCropper);